Move non-public APIs to individual Helper classes
Created helper classes EventHelper, HttpHelper, InitHelper, MainHelper,
ServiceWorkerHelper, and SubscriptionHelper.

As the code improves, some of these classes may go away, but for now the
code has been reorganized so that anything in OneSignal.ts must be a public
API.

Note: Although there are no TypeScript compiler errors, the SDK is currently
broken since some of the .JS files (e.g. notify button / popover) aren't
checked at transpile-time by TypeScript. All files have to be moved over to
TypeScript first.
